# src/app/constants/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is document API for UltraSync",
    "version": "1.0.0",
    "title": "Swagger UltraSync"
  },
  "host": "localhost:3000/",
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/api/event/{event_id}": {
      "put": {
        "tags": ["Events"],
        "summary": "Update an event",
        "description": "",
        "operationId": "updateEvent",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event that need to be edited",
            "schema": {
              "$ref": "#/definitions/eventUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/event"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": ["Events"],
        "summary": "Get Event Detail",
        "description": "",
        "operationId": "eventDetail",
        "produces": ["application/json", "application/xml"],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/eventDetail"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/impact/{impact_id}": {
      "put": {
        "tags": ["Impacts"],
        "summary": "Update Impact",
        "description": "",
        "operationId": "updateImpact",
        "produces": ["application/json", "application/xml"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event that need to be changed",
            "schema": {
              "$ref": "#/definitions/impactUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/impactUpdate"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "eventDetail": {
      "type": "object",
      "properties": {
        "event_id": {
          "type": "integer",
          "example": 1
        },
        "event_name": {
          "type": "string",
          "example": "Manchester United VS Chelsea"
        },
        "event_start_time": {
          "type": "timestamp",
          "example": "2020-20-03:19:00"
        },
        "event_end_time": {
          "type": "timestamp",
          "example": "2020-20-03:21:00"
        },
        "event_type": {
          "type": "integer",
          "example": 1
        },
        "event_state": {
          "type": "string",
          "example": "---"
        },
        "isactive": {
          "type": "boolean",
          "example": "true"
        },
        "ismultiday": {
          "type": "boolean",
          "example": "true"
        },
        "created_by": {
          "type": "string",
          "example": "Admin"
        },
        "updated_by": {
          "type": "string",
          "example": "null"
        },
        "created_at": {
          "type": "timestamp",
          "example": "2020-20-03:19:00"
        },
        "updated_at": {
          "type": "timestamp",
          "example": "null"
        },
        "event_sport_group": {
          "type": "object",
          "properties": {
            "event_sport_group_id": {
              "type": "integer",
              "example": 1
            },
            "event_sport_group_name": {
              "type": "string",
              "example": "MU"
            }
          }
        },
        "tribe": {
          "type": "object",
          "properties": {
            "tribe_id": {
              "type": "integer",
              "example": 1
            },
            "tribe_name": {
              "type": "string",
              "example": "MU"
            },
            "tribe_image_url": {
              "type": "string",
              "example": "URL SAMPLE"
            }
          }
        },
        "opponent_tribe": {
          "type": "object",
          "properties": {
            "tribe_id": {
              "type": "integer",
              "example": 1
            },
            "tribe_name": {
              "type": "string",
              "example": "Chelsea"
            },
            "tribe_image_url": {
              "type": "string",
              "example": "URL SAMPLE"
            }
          }
        },
        "session": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "session_id": {
                "type": "integer",
                "example": 1
              },
              "session_name": {
                "type": "string",
                "example": "Prematch"
              },
              "sound_id": {
                "type": "integer",
                "example": 1
              },
              "sponsor_id": {
                "type": "integer",
                "example": 1
              },
              "sponsor_name": {
                "type": "string",
                "example": "LG"
              },
              "session_duration": {
                "type": "integer",
                "example": 30
              },
              "session_start_time": {
                "type": "timestamp",
                "example": "2020-07-11T02:00:16.505Z"
              },
              "session_max_donation_total": {
                "type": "number",
                "example": 200.341
              },
              "session_max_donation_per_user": {
                "type": "number",
                "example": 0.5
              },
              "session_show_logo": {
                "type": "boolean",
                "example": true
              },
              "session_ads_start": {
                "type": "boolean",
                "example": true
              },
              "session_ads_end": {
                "type": "boolean",
                "example": true
              },
              "language_id": {
                "type": "integer",
                "example": 1
              },
              "language_name": {
                "type": "string",
                "example": "English"
              },
              "session_keyword": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "session_keyword_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "session_keyword_name": {
                      "type": "string",
                      "example": "Go"
                    },
                    "language_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "language_name": {
                      "type": "string",
                      "example": "English"
                    }
                  }
                }
              },
              "impact": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "impact_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "impact_name": {
                      "type": "string",
                      "example": "For Our Children"
                    },
                    "impact_country_name": {
                      "type": "string",
                      "example": "Cambodia"
                    },
                    "impact_category": {
                      "type": "string",
                      "example": "Health"
                    },
                    "language_id": {
                      "type": "integer",
                      "example": 1
                    },
                    "language_name": {
                      "type": "string",
                      "example": "English"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "event": {
      "type": "object",
      "properties": {
        "event_name": {
          "type": "string",
          "example": "Manchester United VS Chelsea"
        },
        "event_description": {
          "type": "string",
          "example": " A Premier League's Football Match"
        },
        "event_address": {
          "type": "string",
          "example": " TMA Address"
        },
        "event_sport_group_id": {
          "type": "integer",
          "example": 1
        },
        "tribe_id": {
          "type": "integer",
          "example": 1
        },
        "opponent_tribe_id": {
          "type": "integer",
          "example": 2
        },
        "event_start_time": {
          "type": "timestamp",
          "example": "2020-20-03:19:00"
        },
        "event_end_time": {
          "type": "timestamp",
          "example": "2020-20-03:21:00"
        },
        "event_type": {
          "type": "integer",
          "example": 1
        },
        "event_repeat": {
          "type": "boolean",
          "example": false
        },
        "event_image_url": {
          "type": "string",
          "example": "Image URL"
        },
        "ismultiday": {
          "type": "boolean",
          "example": false
        },
        "isactive": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "eventUpdate": {
      "type": "object",
      "properties": {
        "event_name": {
          "type": "string",
          "example": "Manchester United VS Chelsea"
        },
        "event_description": {
          "type": "string",
          "example": " A Premier League's Football Match"
        },
        "event_address": {
          "type": "string",
          "example": " TMA Address"
        },
        "event_sport_group_id": {
          "type": "integer",
          "example": 1
        },
        "tribe_id": {
          "type": "integer",
          "example": 1
        },
        "opponent_tribe_id": {
          "type": "integer",
          "example": 2
        },
        "event_start_time": {
          "type": "timestamp",
          "example": "2020-20-03:19:00"
        },
        "event_end_time": {
          "type": "timestamp",
          "example": "2020-20-03:21:00"
        },
        "event_type": {
          "type": "integer",
          "example": 1
        },
        "event_repeat": {
          "type": "boolean",
          "example": false
        },
        "event_image_url": {
          "type": "string",
          "example": "Image URL"
        },
        "ismultiday": {
          "type": "boolean",
          "example": false
        }
      }
    },
...
    "impactUpdate": {
      "type": "object",
      "properties": {
        "impact_name": {
          "type": "string",
          "example": "Prevent Cancer"
        },
        "impact_description": {
          "type": "string",
          "example": "We aim for a better world"
        },
        "impact_achieve": {
          "type": "string",
          "example": "1st nobel peace"
        },
        "impact_total_donation": {
          "type": "number",
          "example": 300.321
        },
        "impact_type": {
          "type": "integer",
          "example": 1
        },
        "impact_image_url": {
          "type": "string",
          "example": "Sample URL"
        },
        "category_id": {
          "type": "integer",
          "example": 2
        },
        "country_id": {
          "type": "integer",
          "example": 33
        },
        "language_id": {
          "type": "integer",
          "example": 1
        },
        "isactive": {
          "type": "boolean",
          "example": true
        }
      }
    },
...
  }
}